WIP commit: fix tests

+++ src/test/java/com/sharethrough/sdk/AdImageViewTest.java
@@ -49,7 +49,7 @@ public class AdImageViewTest extends TestBase {
         creative.renderedTime = Integer.MIN_VALUE;
         creative.wasVisible = true;
 
-        when(sharethrough.getAdCacheTimeInMilliseconds()).thenReturn(20);
+        when(sharethrough.getAdCacheTimeInMilliseconds()).thenReturn((long) 20);
         subject.onAttachedToWindow();
         verify(sharethrough).putCreativeIntoAdView(adView, feedPosition);
     }
@@ -58,7 +58,7 @@ public class AdImageViewTest extends TestBase {
     public void WhenAdImageViewIsAttached_VerifyThatAdViewTimerTaskIsScheduled() throws Exception {
         creative.renderedTime = Integer.MAX_VALUE;
 
-        when(sharethrough.getAdCacheTimeInMilliseconds()).thenReturn(20);
+        when(sharethrough.getAdCacheTimeInMilliseconds()).thenReturn((long) 20);
         subject.onAttachedToWindow();
         verify(timer).schedule(any(TimerTask.class), anyInt(), anyInt());
     }
